+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   Use UnleashConfig.Builder to get instance of UnleashConfig
 - feat: Add annotations to indicate null and nonnull method signatures, also supports Kotlin
 - feat: Added 'more' method to Unleash API to place advanced usecases like evaluating all toggles at once and manually incrementing usage counts
+- feat: Add warning when a matching strategy for a toggle can't be found
 
 ## 3.3.4
 - fix: follow redirect once (#115)